ci: split node.test.mjs into multiple tests to increase pipeline stability
