+++ .github/workflows/zxc-code-analysis.yaml
@@ -65,11 +65,16 @@ on:
         type: string
         required: false
         default: "e2e-mirror-node"
-      e2e-node-test-subdir:
-        description: "E2E Node Test Subdirectory:"
+      e2e-node-pem-stop-add-test-subdir:
+        description: "E2E Node PEM Stop Add Test Subdirectory:"
         type: string
         required: false
-        default: "e2e-node"
+        default: "e2e-node-pem-stop-add"
+      e2e-node-test-subdir:
+        description: "E2E Node PFX Kill Add Test Subdirectory:"
+        type: string
+        required: false
+        default: "e2e-node-pfx-kill-add"
       e2e-relay-test-subdir:
         description: "E2E Relay Test Subdirectory:"
         type: string
@@ -85,11 +90,16 @@ on:
         type: string
         required: false
         default: "E2E Mirror Node Tests Coverage Report"
-      e2e-node-coverage-report:
-        description: "E2E Node Coverage Report:"
+      e2e-node-pem-stop-add-coverage-report:
+        description: "E2E Node PEM Stop Add Coverage Report:"
         type: string
         required: false
-        default: "E2E Node Tests Coverage Report"
+        default: "E2E Node PEM Stop Add Tests Coverage Report"
+      e2e-node-pfx-kill-add-coverage-report:
+        description: "E2E Node PFX Kill Add Coverage Report:"
+        type: string
+        required: false
+        default: "E2E Node PFX Kill Add Tests Coverage Report"
       e2e-relay-coverage-report:
         description: "E2E Relay Coverage Report:"
         type: string
@@ -159,12 +169,19 @@ jobs:
           name: ${{ inputs.e2e-mirror-node-coverage-report }}
           path: 'coverage/${{ inputs.e2e-mirror-node-test-subdir }}'
 
-      - name: Download E2E Node Coverage Report
+      - name: Download E2E Node PEM Stop Add Coverage Report
         uses: actions/download-artifact@65a9edc5881444af0b9093a5e628f2fe47ea3b2e # v4.1.7
         if: ${{ (inputs.enable-codecov-analysis || inputs.enable-codacy-coverage) && inputs.enable-e2e-coverage-report && !cancelled() && !failure() }}
         with:
-          name: ${{ inputs.e2e-node-coverage-report }}
-          path: 'coverage/${{ inputs.e2e-node-test-subdir }}'
+          name: ${{ inputs.e2e-node-pem-stop-add-coverage-report }}
+          path: 'coverage/${{ inputs.e2e-node-pem-stop-add-test-subdir }}'
+
+      - name: Download E2E Node PFX Kill Add Coverage Report
+        uses: actions/download-artifact@65a9edc5881444af0b9093a5e628f2fe47ea3b2e # v4.1.7
+        if: ${{ (inputs.enable-codecov-analysis || inputs.enable-codacy-coverage) && inputs.enable-e2e-coverage-report && !cancelled() && !failure() }}
+        with:
+          name: ${{ inputs.e2e-node-pfx-kill-add-coverage-report }}
+          path: 'coverage/${{ inputs.e2e-node-pfx-kill-add-test-subdir }}'
 
       - name: Download E2E Relay Coverage Report
         uses: actions/download-artifact@65a9edc5881444af0b9093a5e628f2fe47ea3b2e # v4.1.7
